(PC-31176) refactor(booking): extract withSmalBadge in a specific component for more reusability
