+++ src/features/bookings/components/SmallBadgedButton.tsx
@@ -1,30 +1,6 @@
-import React, { ComponentProps, FunctionComponent } from 'react'
-import { View } from 'react-native'
-import styled from 'styled-components/native'
+import { ComponentProps } from 'react'
 
 import { RoundedButton } from 'ui/components/buttons/RoundedButton'
-import { Pastille } from 'ui/svg/icons/Pastille'
+import { WithSmallBadge } from 'ui/components/WithSmallBadge'
 
-const SmallBadge = styled(Pastille).attrs(({ theme }) => ({
-  color: theme.colors.primary,
-  width: 8,
-  height: 8,
-  testID: 'smallBadge',
-}))({
-  position: 'absolute',
-  top: 2,
-  right: 2,
-})
-
-function withSmallBadge<P extends object>(Component: FunctionComponent<P>) {
-  return function ComponentWithSmallBadge(props: P) {
-    return (
-      <View>
-        <Component {...props} />
-        <SmallBadge />
-      </View>
-    )
-  }
-}
-
-export const SmallBadgedButton = withSmallBadge<ComponentProps<typeof RoundedButton>>(RoundedButton)
+export const SmallBadgedButton = WithSmallBadge<ComponentProps<typeof RoundedButton>>(RoundedButton)
